feat: Migrate requestMatcher to TS
